+++ static/js/containers/admin/AdminPage.js
@@ -1,13 +1,12 @@
 // @flow
 import React from "react"
 import { Route } from "react-router-dom"
-import R from "ramda"
 
 import CreateChannelPage from "./CreateChannelPage"
 import EditChannelAppearancePage from "./EditChannelAppearancePage"
 import EditChannelBasicPage from "./EditChannelBasicPage"
-
-import withMemberForm from "../../hoc/withMemberForm"
+import EditChannelModeratorsPage from "./EditChannelModeratorsPage"
+import EditChannelContributorsPage from "./EditChannelContributorsPage"
 
 import type { Match } from "react-router"
 
@@ -31,19 +30,11 @@ export default class AdminPage extends React.Component<Props> {
         />
         <Route
           path={`${match.url}/c/edit/:channelName/members/moderators`}
-          component={withMemberForm(
-            "moderator",
-            "channelModerators",
-            R.prop("moderator_name")
-          )}
+          component={EditChannelModeratorsPage}
         />
         <Route
           path={`${match.url}/c/edit/:channelName/members/contributors`}
-          component={withMemberForm(
-            "contributor",
-            "channelContributors",
-            R.prop("contributor_name")
-          )}
+          component={EditChannelContributorsPage}
         />
       </React.Fragment>
     )
