+++ static/js/hoc/withMemberForm.js
@@ -1,25 +1,75 @@
 // @flow
 import React from "react"
+import R from "ramda"
+import { MetaTags } from "react-meta-tags"
 
-import EditChannelMembersPage from "../containers/admin/EditChannelMembersPage"
+import Card from "../components/Card"
+import EditChannelMembersForm from "../components/admin/EditChannelMembersForm"
+import EditChannelNavbar from "../components/admin/EditChannelNavbar"
 
-import type { Member } from "../flow/discussionTypes"
+import { formatTitle } from "../lib/title"
 
-const withMemberForm = (
-  noun: string,
-  reducerName: string,
-  usernameGetter: (member: Member) => string
-) => {
-  class withMemberForm extends React.Component<*> {
-    render() {
+import type { Channel, Member } from "../flow/discussionTypes"
+
+const shouldLoadData = R.complement(R.allPass([R.eqProps("channelName")]))
+
+const withMemberForm = (WrappedComponent: *) => {
+  type Props = {
+    channel: Channel,
+    loadChannel: () => Promise<*>,
+    loadMembers: () => Promise<*>,
+    members: Array<Member>,
+    usernameGetter: (member: Member) => string
+  }
+
+  class withMemberForm extends React.Component<Props> {
+    componentDidMount() {
+      this.loadData()
+    }
+
+    componentDidUpdate(prevProps: Props) {
+      if (shouldLoadData(prevProps, this.props)) {
+        this.loadData()
+      }
+    }
+
+    loadData = async () => {
+      const { channel, loadChannel, loadMembers, members } = this.props
+      if (!channel) {
+        await loadChannel()
+      }
+
+      if (!members) {
+        await loadMembers()
+      }
+    }
+
+    renderBody = () => {
+      const { channel, members, usernameGetter } = this.props
+
+      if (!channel || !members) {
+        return null
+      }
+
       return (
-        <EditChannelMembersPage
-          noun={noun}
-          reducerName={reducerName}
-          usernameGetter={usernameGetter}
-          {...this.props}
-        />
+        <React.Fragment>
+          <MetaTags>
+            <title>{formatTitle("Edit Channel")}</title>
+          </MetaTags>
+          <EditChannelNavbar channelName={channel.name} />
+          <Card>
+            <EditChannelMembersForm
+              channelName={channel.name}
+              members={members}
+              usernameGetter={usernameGetter}
+            />
+          </Card>
+        </React.Fragment>
       )
+    }
+
+    render() {
+      return <WrappedComponent {...this.props} renderBody={this.renderBody} />
     }
   }
   return withMemberForm
